+++ src/main/java/nextstep/authentication/AuthConfig.java
@@ -1,7 +1,7 @@
-package nextstep.member;
+package nextstep.authentication;
 
-import nextstep.member.application.JwtTokenProvider;
-import nextstep.member.ui.AuthenticationPrincipalArgumentResolver;
+import nextstep.authentication.application.JwtTokenProvider;
+import nextstep.authentication.ui.AuthenticationPrincipalArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
