+++ src/main/java/nextstep/authentication/application/dto/GithubCodeRequest.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.authentication.application.dto;
 
 public class GithubCodeRequest {
     private String code;
