+++ src/main/java/nextstep/authentication/application/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.authentication.application.dto;
 
 public class TokenRequest {
     private String email;
