+++ src/main/java/nextstep/authentication/application/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.member;
+package nextstep.authentication.application.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
