+++ src/main/java/nextstep/authentication/domain/LoginMember.java
@@ -1,4 +1,4 @@
-package nextstep.member.domain;
+package nextstep.authentication.domain;
 
 public class LoginMember {
     private String email;
